mergeSort
