--- challenge10/app/src/main/java/challenge10/AnimalShelter/Animal.java
@@ -1,12 +1,0 @@
-package challenge10.AnimalShelter;
-
-import challenge10.queue.data.QueueNode;
-
-public class Animal<T> extends QueueNode<T> {
-
-    public Animal(T animalName) {
-        super(animalName);
-    }
-
-
-}
