--- challenge10/app/src/main/java/challenge10/AnimalShelter/Cat.java
@@ -1,8 +1,0 @@
-package challenge10.AnimalShelter;
-
-public class Cat<T> extends Animal<T>{
-
-    public Cat(T animalName) {
-        super(animalName);
-    }
-}
