--- challenge10/app/src/main/java/challenge10/AnimalShelter/Dog.java
@@ -1,9 +1,0 @@
-package challenge10.AnimalShelter;
-
-public class Dog<T> extends Animal<T>{
-
-
-    public Dog(T animalName) {
-        super(animalName);
-    }
-}
