--- challenge10/app/src/main/java/challenge10/Trees/data/BinaryNode.java
@@ -1,12 +1,0 @@
-package challenge10.Trees.data;
-
-public class BinaryNode <T extends Comparable<T>> extends BTNode<T>{
-    public BinaryNode(T value) {
-        super(value);
-    }
-
-    @Override
-    public int compareTo(BTNode<T> o) {
-        return 0;
-    }
-}
