+++ challenge10/app/src/main/java/challenges/AnimalShelter/Animal.java
@@ -1,0 +1,12 @@
+package challenges.AnimalShelter;
+
+import challenges.queue.data.QueueNode;
+
+public class Animal<T> extends QueueNode<T> {
+
+    public Animal(T animalName) {
+        super(animalName);
+    }
+
+
+}
