+++ challenge10/app/src/main/java/challenges/AnimalShelter/Cat.java
@@ -1,0 +1,8 @@
+package challenges.AnimalShelter;
+
+public class Cat<T> extends Animal<T>{
+
+    public Cat(T animalName) {
+        super(animalName);
+    }
+}
