+++ challenge10/app/src/main/java/challenges/AnimalShelter/Dog.java
@@ -1,0 +1,9 @@
+package challenges.AnimalShelter;
+
+public class Dog<T> extends Animal<T>{
+
+
+    public Dog(T animalName) {
+        super(animalName);
+    }
+}
