+++ challenge10/app/src/test/java/challenges/AppTest.java
@@ -1,0 +1,498 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package challenges;
+
+import challenges.AnimalShelter.AnimalShelter;
+import challenges.AnimalShelter.Cat;
+import challenges.AnimalShelter.Dog;
+import challenges.InsertionSort.InsertionSort;
+import challenges.MergeSort.MergeSort;
+import challenges.Trees.data.BTNode;
+import challenges.Trees.structure.BinarySearchTree;
+import challenges.Trees.structure.BinaryTree;
+import challenges.pseudoQueue.pseudoQueue;
+import challenges.queue.structure.Queue;
+import challenges.stack.structure.Stack;
+import challenges.validateBrackets.ValidateBrackets;
+import org.junit.jupiter.api.Test;
+
+import java.util.Arrays;
+
+import static challenges.Trees.structure.BinaryTree.TraversalOrder.*;
+import static org.junit.jupiter.api.Assertions.*;
+
+class AppTest {
+    @Test void appHasAGreeting() {
+        App classUnderTest = new App();
+        assertNotNull(classUnderTest.getGreeting(), "app should have a greeting");
+    }
+    @Test void stackPushTest(){
+        Stack newStack=new Stack();
+        newStack.push("1");
+        assertEquals(1,newStack.getSize());
+    }
+    @Test void stackMultiplePushTest(){
+        Stack newStack=new Stack();
+        newStack.push("1");
+        newStack.push("2");
+        newStack.push("3");
+        newStack.push("4");
+        newStack.push("5");
+        assertEquals(5,newStack.getSize());
+    }
+    @Test void stackPopTest() throws Exception {
+        Stack newStack=new Stack();
+        newStack.push("1");
+        newStack.push("2");
+        newStack.push("3");
+        newStack.push("4");
+        newStack.push("5");
+        System.out.println(newStack.pop());
+        assertEquals(4,newStack.getSize());
+    }
+    @Test void stackMultiplePopTest() throws Exception {
+        Stack newStack=new Stack();
+        newStack.push("1");
+        newStack.push("2");
+        newStack.push("3");
+        newStack.push("4");
+        newStack.push("5");
+        System.out.println(newStack.pop());
+        System.out.println(newStack.pop());
+        System.out.println(newStack.pop());
+        System.out.println(newStack.pop());
+        System.out.println(newStack.pop());
+        assertEquals(0,newStack.getSize());
+    }
+    @Test void stackPeekTest() throws Exception {
+        Stack newStack=new Stack();
+        newStack.push("1");
+        newStack.push("2");
+        newStack.push("3");
+        newStack.push("4");
+        newStack.push("5");
+        System.out.println(newStack.peek());
+        assertEquals("5",newStack.peek());
+    }
+    @Test void stackEmptyTest(){
+        Stack newStack=new Stack();
+        assertNull(newStack.getTop());
+    }
+    @Test void stackEmptyPopTest()  {
+        assertThrows(Exception.class, () -> {
+            Stack newStack=new Stack();
+            newStack.pop();
+        });
+    }
+    @Test void stackEmptyPeekTest(){
+        assertThrows(Exception.class, () -> {
+            Stack newStack=new Stack();
+            String val=newStack.peek();
+        });
+    }
+    @Test void queueEnqueueTest(){
+        Queue<String> newQueue=new Queue<String>();
+        newQueue.enqueue("1");
+        assertEquals(1,newQueue.getSize());
+    }
+    @Test void queueEnqueueMultipleTest(){
+        Queue<String> newQueue=new Queue<>();
+        newQueue.enqueue("1");
+        newQueue.enqueue("2");
+        newQueue.enqueue("3");
+        newQueue.enqueue("4");
+        newQueue.enqueue("5");
+        assertEquals(5,newQueue.getSize());
+    }
+    @Test void DequeueTest() throws Exception {
+        Queue<String> newQueue=new Queue<String>();
+        newQueue.enqueue("1");
+        newQueue.enqueue("2");
+        newQueue.enqueue("3");
+        newQueue.enqueue("4");
+        newQueue.enqueue("5");
+        System.out.println(newQueue.dequeue());
+        assertEquals(4,newQueue.getSize());
+    }
+    @Test void queuePeekTest() throws Exception {
+        Queue<String> newQueue=new Queue<>();
+        newQueue.enqueue("1");
+        newQueue.enqueue("2");
+        newQueue.enqueue("3");
+        newQueue.enqueue("4");
+        newQueue.enqueue("5");
+        System.out.println(newQueue.peek());
+        assertEquals("1",newQueue.peek());
+    }
+    @Test void DequeueMultipleTest() throws Exception {
+        Queue<String> newQueue=new Queue<String>();
+        newQueue.enqueue("1");
+        newQueue.enqueue("2");
+        newQueue.enqueue("3");
+        newQueue.enqueue("4");
+        newQueue.enqueue("5");
+        System.out.println(newQueue.dequeue());
+        System.out.println(newQueue.dequeue());
+        System.out.println(newQueue.dequeue());
+        System.out.println(newQueue.dequeue());
+        System.out.println(newQueue.dequeue());
+        assertEquals(0,newQueue.getSize());
+    }
+    @Test void EmptyQueueTest(){
+        Queue<String> newQueue=new Queue<>();
+        assertNull(newQueue.getFront());
+    }
+    @Test void queueEmptyPeekTest(){
+        assertThrows(Exception.class, () -> {
+            Queue<String> newQueue=new Queue<>();
+            String val=newQueue.peek();
+            System.out.println(val);
+        });
+    }
+    @Test void queueEmptyDequeueTest(){
+        assertThrows(Exception.class, () -> {
+            Queue<String> newQueue=new Queue<>();
+            String val=newQueue.dequeue();
+            System.out.println(val);
+        });
+    }
+    @Test void pseudoQueueTest() throws Exception {
+        pseudoQueue newQueue=new pseudoQueue();
+        newQueue.enQueue("1");
+        newQueue.enQueue("2");
+        newQueue.enQueue("3");
+        newQueue.enQueue("4");
+
+        assertEquals("1",newQueue.getS1().getTop().getValue());
+    }
+    @Test void pseudoDequeueTest() throws Exception {
+        pseudoQueue newQueue=new pseudoQueue();
+        newQueue.enQueue("1");
+        newQueue.enQueue("2");
+        newQueue.enQueue("3");
+        newQueue.enQueue("4");
+        String last=newQueue.deQueue();
+        assertEquals("1",last);
+    }
+
+    @Test void animalShelterTest() throws Exception {
+        AnimalShelter<String> animal=new AnimalShelter<>();
+        animal.enQueue(new Cat<>("cat1"));
+        animal.enQueue(new Dog<>("dog1"));
+        animal.enQueue(new Cat<>("cat2"));
+        animal.enQueue(new Dog<>("dog2"));
+        animal.enQueue(new Cat<>("cat3"));
+
+        assertEquals("cat1",animal.deQueue("cat").getValue());
+    }
+    @Test void animalShelterSizeTest() throws Exception {
+        AnimalShelter<String> animal=new AnimalShelter<>();
+        animal.enQueue(new Cat<>("cat1"));
+        animal.enQueue(new Dog<>("dog1"));
+        animal.enQueue(new Cat<>("cat2"));
+        animal.enQueue(new Dog<>("dog2"));
+        animal.enQueue(new Cat<>("cat3"));
+
+        assertEquals(5,animal.getSize());
+    }
+    @Test void animalShelterEmptyTest() throws Exception {
+        AnimalShelter<String> animal=new AnimalShelter<>();
+        animal.enQueue(new Dog<>("dog1"));
+        animal.enQueue(new Dog<>("dog2"));
+
+        assertNull(animal.deQueue("cat"));
+    }
+    @Test void validateBracketsTrueTest(){
+        String str="{}{Code}[Fellows](())";
+        ValidateBrackets ob=new ValidateBrackets();
+        assertEquals(true,ob.validateBrackets(str));
+    }
+    @Test void validateBracketsTrue2Test(){
+        String str="{}";
+        ValidateBrackets ob=new ValidateBrackets();
+        assertEquals(true,ob.validateBrackets(str));
+    }
+    @Test void validateBracketsTrue3Test(){
+        String str="{}(){}";
+        ValidateBrackets ob=new ValidateBrackets();
+        assertEquals(true,ob.validateBrackets(str));
+    }
+    @Test void validateBracketsTrue4Test(){
+        String str="()[[Extra Characters]]";
+        ValidateBrackets ob=new ValidateBrackets();
+        assertEquals(true,ob.validateBrackets(str));
+    }
+    @Test void validateBracketsTrue5Test(){
+        String str="(){}[[]]";
+        ValidateBrackets ob=new ValidateBrackets();
+        assertEquals(true,ob.validateBrackets(str));
+    }
+    @Test void validateBracketsFalse1Test(){
+        String str="[({}]";
+        ValidateBrackets ob=new ValidateBrackets();
+        assertEquals(false,ob.validateBrackets(str));
+    }
+    @Test void validateBracketsFalse2Test(){
+        String str=")";
+        ValidateBrackets ob=new ValidateBrackets();
+        assertEquals(false,ob.validateBrackets(str));
+    }
+    @Test void validateBracketsFalse3Test(){
+        String str="(](";
+        ValidateBrackets ob=new ValidateBrackets();
+        assertEquals(false,ob.validateBrackets(str));
+    }
+    @Test void validateBracketsFalse4Test(){
+        String str="{(})";
+        ValidateBrackets ob=new ValidateBrackets();
+        assertEquals(false,ob.validateBrackets(str));
+    }
+    @Test void validateBracketsFalse5Test(){
+        String str="{";
+        ValidateBrackets ob=new ValidateBrackets();
+        assertEquals(false,ob.validateBrackets(str));
+    }
+    @Test void validateBracketsFalse6Test(){
+        String str="[}";
+        ValidateBrackets ob=new ValidateBrackets();
+        assertEquals(false,ob.validateBrackets(str));
+    }
+
+    @Test void EmptyBinaryTreeTest(){
+        BinaryTree<Integer> tre=new BinaryTree<>();
+        assertNull(tre.getRoot());
+    }
+    @Test void BinaryTreeTest(){
+        BinaryTree<Integer> tre=new BinaryTree<>(9);
+        assertNotNull(tre.getRoot());
+    }
+    @Test void BinarySearchTreeLeftTest(){
+        BinarySearchTree<Integer> newTree=new BinarySearchTree<>();
+        newTree.add(5);
+        newTree.add(8);
+        newTree.add(4);
+        assertEquals(4,newTree.getRoot().getLeftNode().getData());
+    }
+    @Test void BinarySearchTreeRightTest(){
+        BinarySearchTree<Integer> newTree=new BinarySearchTree<>();
+        newTree.add(5);
+        newTree.add(8);
+        newTree.add(4);
+        assertEquals(8,newTree.getRoot().getRightNode().getData());
+    }
+    @Test void BinarySearchTreeRootTest(){
+        BinarySearchTree<Integer> newTree=new BinarySearchTree<>();
+        newTree.add(5);
+        newTree.add(8);
+        newTree.add(4);
+        assertEquals(5,newTree.getRoot().getData());
+    }
+    @Test void BinarySearchTreeAddTest(){
+        BinarySearchTree<Integer> newTree=new BinarySearchTree<>();
+        newTree.add(5);
+        newTree.getRoot().setLeftNode(new BTNode<Integer>(8) {
+            @Override
+            public int compareTo(BTNode<Integer> o) {
+                return 0;
+            }
+        });
+        newTree.getRoot().setRightNode(new BTNode<Integer>(4) {
+            @Override
+            public int compareTo(BTNode<Integer> o) {
+                return 0;
+            }
+        });
+        assertEquals(8,newTree.getRoot().getLeftNode().getData());
+    }
+    @Test void BinarySearchTreePreorderTest(){
+        BinarySearchTree<Integer> newTree=new BinarySearchTree<>();
+        newTree.add(5);
+        newTree.add(8);
+        newTree.add(4);
+        newTree.add(3);
+        newTree.add(4);
+        newTree.add(90);
+        newTree.add(55);
+        assertEquals("[5, 4, 3, 8, 90, 55]",newTree.traverse(PREORDER).toString());
+    }
+    @Test void BinarySearchTreeInorderTest(){
+        BinarySearchTree<Integer> newTree=new BinarySearchTree<>();
+        newTree.add(5);
+        newTree.add(8);
+        newTree.add(4);
+        newTree.add(3);
+        newTree.add(4);
+        newTree.add(90);
+        newTree.add(55);
+        assertEquals("[3, 4, 5, 8, 55, 90]",newTree.traverse(INORDER).toString());
+    }
+    @Test void BinarySearchTreePostorderTest(){
+        BinarySearchTree<Integer> newTree=new BinarySearchTree<>();
+        newTree.add(5);
+        newTree.add(8);
+        newTree.add(4);
+        newTree.add(3);
+        newTree.add(4);
+        newTree.add(90);
+        newTree.add(55);
+        assertEquals("[3, 4, 55, 90, 8, 5]",newTree.traverse(POSTORDER).toString());
+    }
+    @Test void BinarySearchTreeContainsTrueTest(){
+        BinarySearchTree<Integer> newTree=new BinarySearchTree<>();
+        newTree.add(5);
+        newTree.add(8);
+        newTree.add(4);
+        newTree.add(3);
+        newTree.add(4);
+        newTree.add(90);
+        newTree.add(55);
+        assertTrue(newTree.Contains(90));
+    }
+    @Test void BinarySearchTreeContainsFalseTest(){
+        BinarySearchTree<Integer> newTree=new BinarySearchTree<>();
+        newTree.add(5);
+        newTree.add(8);
+        newTree.add(4);
+        newTree.add(3);
+        newTree.add(4);
+        newTree.add(90);
+        newTree.add(55);
+        assertFalse(newTree.Contains(9));
+    }
+    @Test void BinarySearchTreeSizeTest(){
+        BinarySearchTree<Integer> newTree=new BinarySearchTree<>();
+        newTree.add(5);
+        newTree.add(8);
+        newTree.add(4);
+        newTree.add(3);
+        newTree.add(4);
+        newTree.add(90);
+        newTree.add(55);
+        assertEquals(6,newTree.getSize()); //it gives me 6 because binary search tree doesn't add redundant value
+    }
+
+    @Test void breadthFirstTest() throws Exception {
+        BinarySearchTree<Integer> newTree=new BinarySearchTree<>();
+        newTree.add(5);
+        newTree.add(8);
+        newTree.add(4);
+        newTree.add(3);
+        newTree.add(4);
+        newTree.add(90);
+        newTree.add(55);
+        assertEquals("[5, 4, 8, 3, 90, 55]",newTree.breadthFirst().toString());
+    }
+    @Test void breadthFirstEmptyTest() throws Exception {
+        BinarySearchTree<Integer> newTree=new BinarySearchTree<>();
+
+        assertNull(newTree.breadthFirst());
+    }
+    @Test void maxTreeTest(){
+        BinaryTree<Integer> maxTree=new BinaryTree<Integer>();
+        BTNode<Integer> node1=new BTNode<Integer>(2);
+        maxTree.setRoot(node1);
+
+        node1.setLeftNode(new BTNode<Integer>(7));
+        node1.setRightNode(new BTNode<Integer>(5));
+
+        node1.getLeftNode().setLeftNode(new BTNode<Integer>(2));
+        node1.getLeftNode().setRightNode(new BTNode<Integer>(6));
+
+        node1.getLeftNode().getRightNode().setLeftNode(new BTNode<Integer>(5));
+        node1.getLeftNode().getRightNode().setRightNode(new BTNode<Integer>(11));
+
+        node1.getRightNode().setRightNode(new BTNode<Integer>(9));
+        node1.getRightNode().getRightNode().setLeftNode(new BTNode<Integer>(4));
+
+        assertEquals(11,maxTree.getMaxNum());
+    }
+    @Test void maxTreeEmptyTest(){
+        BinaryTree<Integer> maxTree=new BinaryTree<Integer>();
+        assertNull(maxTree.getMaxNum());
+    }
+    @Test void maxTreeOneNumTest(){
+        BinaryTree<Integer> maxTree=new BinaryTree<Integer>();
+        BTNode<Integer> node1=new BTNode<Integer>(2);
+        maxTree.setRoot(node1);
+        assertEquals(2,maxTree.getMaxNum());
+    }
+
+    @Test void maxTreeNegativeTest(){
+        BinaryTree<Integer> maxTree=new BinaryTree<Integer>();
+        BTNode<Integer> node1=new BTNode<Integer>(-2);
+        maxTree.setRoot(node1);
+
+        node1.setLeftNode(new BTNode<Integer>(-7));
+        node1.setRightNode(new BTNode<Integer>(-5));
+
+        node1.getLeftNode().setLeftNode(new BTNode<Integer>(-2));
+        node1.getLeftNode().setRightNode(new BTNode<Integer>(-6));
+
+        node1.getLeftNode().getRightNode().setLeftNode(new BTNode<Integer>(-5));
+        node1.getLeftNode().getRightNode().setRightNode(new BTNode<Integer>(-11));
+
+        node1.getRightNode().setRightNode(new BTNode<Integer>(-9));
+        node1.getRightNode().getRightNode().setLeftNode(new BTNode<Integer>(-4));
+
+        assertEquals(-2,maxTree.getMaxNum());
+    }
+    @Test void fizzbuzzTest(){
+        BinaryTree<Integer> fizzbuzz=new BinaryTree<>();
+        fizzbuzz.setRoot(new BTNode<>(5));
+        fizzbuzz.getRoot().setLeftNode(new BTNode<>(4));
+        fizzbuzz.getRoot().setRightNode(new BTNode<>(3));
+        fizzbuzz.getRoot().getLeftNode().setLeftNode(new BTNode<>(15));
+        fizzbuzz.getRoot().getLeftNode().setRightNode(new BTNode<>(9));
+        fizzbuzz.getRoot().getRightNode().setLeftNode(new BTNode<>(6));
+        fizzbuzz.getRoot().getRightNode().setRightNode(new BTNode<>(7));
+        System.out.println(fizzbuzz.fizzbuzzTree(fizzbuzz));
+        assertEquals("Buzz",fizzbuzz.fizzbuzzTree(fizzbuzz).getRoot().getData());
+    }
+    @Test void fizzbuzzLeftNodeTest(){
+        BinaryTree<Integer> fizzbuzz=new BinaryTree<>();
+        fizzbuzz.setRoot(new BTNode<>(5));
+        fizzbuzz.getRoot().setLeftNode(new BTNode<>(4));
+        fizzbuzz.getRoot().setRightNode(new BTNode<>(3));
+        fizzbuzz.getRoot().getLeftNode().setLeftNode(new BTNode<>(15));
+        fizzbuzz.getRoot().getLeftNode().setRightNode(new BTNode<>(9));
+        fizzbuzz.getRoot().getRightNode().setLeftNode(new BTNode<>(6));
+        fizzbuzz.getRoot().getRightNode().setRightNode(new BTNode<>(7));
+        System.out.println(fizzbuzz.fizzbuzzTree(fizzbuzz));
+        assertEquals("4",fizzbuzz.fizzbuzzTree(fizzbuzz).getRoot().getLeftNode().getData());
+    }
+    @Test void fizzbuzzNullTest(){
+        BinaryTree<Integer> fizzbuzz=new BinaryTree<>();
+
+        System.out.println(fizzbuzz.fizzbuzzTree(fizzbuzz));
+        assertNull(fizzbuzz.fizzbuzzTree(fizzbuzz).getRoot().getData());
+    }
+    @Test void InsertionTest(){
+        int[] ar={8,4,23,42,16,15};
+        InsertionSort sortedArray=new InsertionSort();
+        System.out.println(Arrays.toString(sortedArray.Insertion(ar)));
+        assertEquals("[4, 8, 15, 16, 23, 42]",Arrays.toString(sortedArray.Insertion(ar)));
+    }
+
+    @Test void MergeTest(){
+        int[] ar2={8,4,23,42,16,15};
+        MergeSort mergeSort=new MergeSort();
+        System.out.println(Arrays.toString(mergeSort.Mergesort(ar2)));
+        assertEquals("[4, 8, 15, 16, 23, 42]",Arrays.toString(mergeSort.Mergesort(ar2)));
+    }
+    @Test void MergeOneTest(){
+        int[] ar2={8};
+        MergeSort mergeSort=new MergeSort();
+        System.out.println(Arrays.toString(mergeSort.Mergesort(ar2)));
+        assertEquals("[8]",Arrays.toString(mergeSort.Mergesort(ar2)));
+    }
+    @Test void MergeEmptyTest(){
+        int[] ar2={};
+        MergeSort mergeSort=new MergeSort();
+        System.out.println(Arrays.toString(mergeSort.Mergesort(ar2)));
+        assertEquals("[]",Arrays.toString(mergeSort.Mergesort(ar2)));
+    }
+
+
+
+}
